Refactor download-logs command
